初始化提交

+++ src/main/java/com/sipc/intelligentfarmbackend/pojo/domain/FieldStatusAlarm.java
@@ -35,12 +35,12 @@ public class FieldStatusAlarm implements Serializable {
     * 
     */
     @TableField("max")
-    private BigDecimal max;
+    private Double max;
     /**
     * 
     */
     @TableField("min")
-    private BigDecimal min;
+    private Double min;
 
 
 }
